+++ eventsource/repository.go
@@ -53,14 +53,6 @@ type repository struct {
 	serializer Serializer
 }
 
-func getTypeOfValue(input interface{}) reflect.Type {
-	value := reflect.TypeOf(input)
-	if value.Kind() == reflect.Ptr {
-		value = value.Elem()
-	}
-	return value
-}
-
 // SaveEvent ...
 func (repo *repository) Save(events ...Event) (err error) {
 	for _, event := range events {
